New support functions for the monster kills script
Some new functions that will now help to use the kills script in a more
secure way. This is done in order to avoid strange crashes with the quest
system or any other script that uses the structures of this script in a
clumsy way.

Also this is done to make Merung stop whining about some entries in the
script error log.