Marked join function as deprecated.
The base.common.join function is a slow version of table.concat(). In call cases the table library function should be used instead of this function. For this reason the join function is marked as deprecated and should be removed as soon as there are no more cases where thsi function is used.