Correct QuestAvailability example
Do not use magic numbers, where possible (e.g. town IDs)
Player.questWillBeAvailable is for quests which you will be able to accept soon,
e.g. if your skill is too low. It is not for active quests or quests that will not be
available under normal circumstances (e.g. wrong faction membership).