Changing script encoding to ISO-8859-1
This conversation was done automatically. It was only performed on
files that do not contain any encoding errors. Files that were not
encoded as UTF-8 remained also untouched. A few files were detected
with the us-ascii encoding, that remain the same, as they do not
contain any characters to distinguish UTF-8 and ISO-8859-1.

Some files contained encoding errors, those need to be fixed by
hand. They got most likely damaged in former attempts to change
the encoding.