updates for evilrock, ferry, northern islands and elstree.
evilrock (0010044)
- selection dialogue will not disappear.
- kept all types of monsters. Pirates of the Caribbean has also undeads
as pirates and this here is fantasy game though. ;)
- numbers of monsters depends now on the numbers of travellers.
- dead players are warped to mainland now in order to avoid being
trapped on the ship.
- portal books will save your ass.
- monsters will be warped to other spot before they get killed by the
script in order to avoid extra loot on the ship.
- players who log out during event is triggered are save now. Either
they will be warped from the ship when they log in afterwards or there
will be a portal that leads them to the main land.
- I could not reproduce the massage that I earn rankpoints but only that
I lose rankpoints. No idea how this happened to you.
- map of the ship was adjusted that the NPC can be seen better.

evilrock
- adjust to background story

northern island and elstree
- multiple chars can trigger effects now