Quick and dirty fix for the multi-hit issue of cone attacks
This fix works around the problem that the rawequal function is not
working as expected for position structures. This results in the problem
that the same position that is part of different structures is handled
as separate index.

This fix works around this issue by checking the already present entries
in the list for the position that is to be added one by one.